api/requirements.txt
flask==1.1.1
pydub==0.23.1
pytube==9.5.2
gensim==3.8.1
google-cloud==0.34.0
google-cloud-speech==1.2.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
google-cloud-pubsub==1.0.2